feat(DTFS2-6508): update feedback access (#2059)
