+++ cspell.json
@@ -1,143 +1,134 @@
 {
-    "ignorePaths": [
-        "**/node_modules/**",
-        "**/vscode-extension/**",
-        "**/.git/**",
-        "**/coverage/**",
-        ".vscode",
-        ".gitignore",
-        "package-lock.json",
-        "package.json",
-        "report",
-        "dist/**",
-        "**/public/**",
-        "**/webpack.common.**",
-        ".gitignore",
-        "**/Dockerfile"
-    ],
-    "language": "en",
-    "noConfigSearch": true,
-    "allowCompoundWords": true,
-    "words": [
-        "Abhi",
-        "acbs",
-        "aChecker",
-        "actionsheet",
-        "ADMINNOMAKER",
-        "APIM",
-        "azurecr",
-        "BBALIBOR",
-        "behaviour",
-        "Bloggs",
-        "bson",
-        "centralised",
-        "CICD",
-        "CILC",
-        "Clickjacking",
-        "codacy",
-        "codecov",
-        "codeql",
-        "commitlintrc",
-        "creds",
-        "crsf",
-        "Curncy",
-        "datafixes",
-        "Dockerised",
-        "DTFS",
-        "dtos",
-        "ecgd",
-        "EESWE3",
-        "enddate",
-        "ESRA",
-        "Estore",
-        "ESTR",
-        "EWCS",
-        "exisitant",
-        "falsey",
-        "fieldname",
-        "financials",
-        "firstname",
-        "globaldb",
-        "govuk",
-        "healthcheck",
-        "HSTS",
-        "huskyrc",
-        "Informatica",
-        "INITDB",
-        "Initialised",
-        "isready",
-        "jspf",
-        "limitter",
-        "lintstagedrc",
-        "localise",
-        "MAKENCHECK",
-        "Markan",
-        "mdm",
-        "MRUI",
-        "Mulesoft",
-        "nestjs",
-        "noopen",
-        "nosniff",
-        "npmrc",
-        "numbergenerator",
-        "NVARCHAR",
-        "ORDANANCE",
-        "organisation",
-        "pino",
-        "pinojs",
-        "plusplus",
-        "productgroup",
-        "Robo",
-        "sanitisation",
-        "Sanitise",
-        "sefntb",
-        "sess",
-        "snet",
-        "ssword",
-        "startdate",
-        "szenius",
-        "typeorm",
-        "Typeorm",
-        "typescript",
-        "ukef",
-        "UKEF",
-        "ukexportfinance",
-        "uksouth",
-        "Unauthorised",
-        "venv",
-        "VNET",
-        "XLXS",
-        "Stringifying",
-        "SSRF",
-        "TYPEA",
-        "tfsdev",
-        "tfsstaging",
-        "tfsprod",
-        "octocat",
-        "ADMINNOMAKER",
-        "MAKENCHECK",
-        "unmock",
-        "unsubmitted",
-        "wordprocessingml"
-    ],
-    "dictionaries": [
-        "en-gb",
-        "companies",
-        "softwareTerms",
-        "misc",
-        "lorem-ipsum",
-        "typescript",
-        "node",
-        "bash",
-        "npm"
-    ],
-    "languageSettings": [
-        {
-            "languageId": "commit-msg",
-            "ignoreRegExpList": [
-                "/^#.*/gm"
-            ]
-        }
-    ],
-    "version": "0.2"
+  "ignorePaths": [
+    "**/node_modules/**",
+    "**/vscode-extension/**",
+    "**/.git/**",
+    "**/coverage/**",
+    ".vscode",
+    ".gitignore",
+    "package-lock.json",
+    "package.json",
+    "report",
+    "dist/**",
+    "**/public/**",
+    "**/webpack.common.**",
+    ".gitignore",
+    "**/Dockerfile"
+  ],
+  "language": "en",
+  "noConfigSearch": true,
+  "allowCompoundWords": true,
+  "words": [
+    "Abhi",
+    "acbs",
+    "aChecker",
+    "actionsheet",
+    "ADMINNOMAKER",
+    "APIM",
+    "azurecr",
+    "BBALIBOR",
+    "behaviour",
+    "Bloggs",
+    "bson",
+    "centralised",
+    "CICD",
+    "CILC",
+    "Clickjacking",
+    "codacy",
+    "codecov",
+    "codeql",
+    "commitlintrc",
+    "creds",
+    "crsf",
+    "Curncy",
+    "datafixes",
+    "Dockerised",
+    "DTFS",
+    "dtos",
+    "ecgd",
+    "EESWE3",
+    "enddate",
+    "ESRA",
+    "Estore",
+    "ESTR",
+    "EWCS",
+    "exisitant",
+    "falsey",
+    "fieldname",
+    "financials",
+    "firstname",
+    "globaldb",
+    "govuk",
+    "healthcheck",
+    "HSTS",
+    "huskyrc",
+    "Informatica",
+    "INITDB",
+    "Initialised",
+    "isready",
+    "jspf",
+    "limitter",
+    "lintstagedrc",
+    "localise",
+    "MAKENCHECK",
+    "Markan",
+    "mdm",
+    "MRUI",
+    "Mulesoft",
+    "nestjs",
+    "nofollow",
+    "noopen",
+    "noopener",
+    "noreferrer",
+    "nosniff",
+    "npmrc",
+    "numbergenerator",
+    "NVARCHAR",
+    "ORDANANCE",
+    "organisation",
+    "pino",
+    "pinojs",
+    "plusplus",
+    "productgroup",
+    "Robo",
+    "sanitisation",
+    "Sanitise",
+    "sefntb",
+    "sess",
+    "snet",
+    "ssword",
+    "startdate",
+    "szenius",
+    "typeorm",
+    "Typeorm",
+    "typescript",
+    "ukef",
+    "UKEF",
+    "ukexportfinance",
+    "uksouth",
+    "Unauthorised",
+    "venv",
+    "VNET",
+    "XLXS",
+    "Stringifying",
+    "SSRF",
+    "TYPEA",
+    "tfsdev",
+    "tfsstaging",
+    "tfsprod",
+    "octocat",
+    "ADMINNOMAKER",
+    "MAKENCHECK",
+    "unmock",
+    "unsubmitted",
+    "wordprocessingml"
+  ],
+  "dictionaries": ["en-gb", "companies", "softwareTerms", "misc", "lorem-ipsum", "typescript", "node", "bash", "npm"],
+  "languageSettings": [
+    {
+      "languageId": "commit-msg",
+      "ignoreRegExpList": ["/^#.*/gm"]
+    }
+  ],
+  "version": "0.2"
 }
