+++ gef-ui/templates/includes/footer.njk
@@ -18,23 +18,21 @@
               Cookies
             </a>
           </li>
-          {% if isChecker or isMaker %}
-            <li class="govuk-footer__inline-list-item">
-              <a class="govuk-footer__link" href="/feedback" data-cy="feedback-link">
-                Feedback
-              </a>
-            </li>
-          {% endif %}
-            <li class="govuk-footer__inline-list-item">
-              <a class="govuk-footer__link" href="/accessibility-statement" data-cy="accessibility-statement-link">
-                Accessibility statement
-              </a>
-            </li>
-            <li class="govuk-footer__inline-list-item">
-              <a class="govuk-footer__link" target="_blank" rel="noreferrer noopener nofollow" href="https://www.gov.uk/guidance/report-a-vulnerability-on-a-ukef-system">
-                Report a vulnerability
-              </a>
-            </li>
+          <li class="govuk-footer__inline-list-item">
+            <a class="govuk-footer__link" href="/feedback" data-cy="feedback-link">
+              Feedback
+            </a>
+          </li>
+          <li class="govuk-footer__inline-list-item">
+            <a class="govuk-footer__link" href="/accessibility-statement" data-cy="accessibility-statement-link">
+              Accessibility statement
+            </a>
+          </li>
+          <li class="govuk-footer__inline-list-item">
+            <a class="govuk-footer__link" target="_blank" rel="noreferrer noopener nofollow" href="https://www.gov.uk/guidance/report-a-vulnerability-on-a-ukef-system">
+              Report a vulnerability
+            </a>
+          </li>
         </ul>
 
         <svg
